Move ABI parsing to separate module, given it should be independent of Vyper
