+++ boa/abi/event.py
@@ -1,5 +1,6 @@
-from vyper.semantics.types import DArrayT, EventT, VyperType
-from vyper.semantics.types.utils import type_from_abi
+from vyper.semantics.types import EventT
+
+from boa.abi.type import parse_abi_type
 
 
 class ABIEvent(EventT):
@@ -22,11 +23,3 @@ class ABIEvent(EventT):
             arguments={items["name"]: parse_abi_type(items) for items in abi["inputs"]},
             indexed=[i["indexed"] for i in abi["inputs"]],
         )
-
-
-def parse_abi_type(abi: dict) -> VyperType:
-    if abi["type"].endswith("[]"):
-        items_type_name = abi["type"].removesuffix("[]")
-        items_type = type_from_abi({"type": items_type_name})
-        return DArrayT(items_type, 2**256 - 1)
-    return type_from_abi(abi)
