+++ boa/vyper/contract.py
@@ -5,12 +5,8 @@
 import contextlib
 import copy
 import warnings
-from collections import Counter
 from dataclasses import dataclass
 from functools import cached_property
-from itertools import groupby
-from operator import attrgetter
-from os.path import basename
 from typing import Any, Optional
 
 import vyper
@@ -18,10 +14,9 @@ import vyper.ast as vy_ast
 import vyper.ir.compile_ir as compile_ir
 import vyper.semantics.analysis as analysis
 import vyper.semantics.namespace as vy_ns
-from eth.abc import ComputationAPI
 from eth.exceptions import VMError
 from vyper.ast.utils import parse_to_ast
-from vyper.codegen.core import anchor_opt_level, calculate_type_for_external_return
+from vyper.codegen.core import anchor_opt_level
 from vyper.codegen.function_definitions import generate_ir_for_function
 from vyper.codegen.function_definitions.common import ExternalFuncIR, InternalFuncIR
 from vyper.codegen.global_context import GlobalContext
@@ -33,22 +28,20 @@ from vyper.compiler.settings import OptimizationLevel
 from vyper.evm.opcodes import anchor_evm_version
 from vyper.exceptions import VyperException
 from vyper.ir.optimizer import optimize
-from vyper.semantics.analysis.base import FunctionVisibility, StateMutability
 from vyper.semantics.analysis.data_positions import set_data_positions
-from vyper.semantics.types import AddressT, EventT, HashMapT, TupleT, VyperType
-from vyper.semantics.types.function import PositionalArg, _generate_method_id
-from vyper.semantics.types.utils import type_from_abi
+from vyper.semantics.types import AddressT, HashMapT, TupleT
 from vyper.utils import method_id
 
+from boa.abi.contract import _EvmContract
+from boa.abi.errors import BoaError, StackTrace, _handle_child_trace
+from boa.abi.function import _EvmFunction
 from boa.environment import Address, Env
-from boa.profiling import LineProfile, cache_gas_used_for_computation
+from boa.profiling import LineProfile
 from boa.util.abi import abi_decode, abi_encode
 from boa.util.exceptions import strip_internal_frames
 from boa.util.lrudict import lrudict
-from boa.vm.gas_meters import ProfilingGasMeter
 from boa.vm.utils import to_bytes, to_int
 from boa.vyper import _METHOD_ID_VAR
-from boa.vyper.abi import ABIEvent, parse_abi_type
 from boa.vyper.ast_utils import ast_map_of, get_fn_ancestor_from_node, reason_at
 from boa.vyper.compiler_utils import (
     anchor_compiler_settings,
@@ -111,58 +104,6 @@ class VyperDeployer:
         ret.env.register_contract(address, ret)
 
         return ret
-
-
-# base class for ABI and Vyper contracts
-class _EvmContract:
-    def __init__(
-        self,
-        env: Optional[Env] = None,
-        filename: Optional[str] = None,
-        address: Optional[Address] = None,
-    ):
-        self.env = env or Env.get_singleton()
-        self._address = address  # this is overridden by subclasses
-        self.filename = filename
-        self._computation: Optional[ComputationAPI] = None
-
-    def stack_trace(self, computation=None):
-        raise NotImplementedError
-
-    def marshal_to_python(self, computation, vyper_typ):
-        self._computation = computation  # for further inspection
-
-        if computation.is_error:
-            self.handle_error(computation)
-
-        # cache gas used for call if profiling is enabled
-        gas_meter = self.env.vm.state.computation_class._gas_meter_class
-        if gas_meter == ProfilingGasMeter:
-            cache_gas_used_for_computation(self, computation)
-
-        if vyper_typ is None:
-            return None
-
-        return_typ = calculate_type_for_external_return(vyper_typ)
-        ret = abi_decode(return_typ.abi_type.selector_name(), computation.output)
-
-        # unwrap the tuple if needed
-        if not isinstance(vyper_typ, TupleT):
-            (ret,) = ret
-
-        return vyper_object(ret, vyper_typ)
-
-    def handle_error(self, computation) -> None:
-        try:
-            raise BoaError(self.stack_trace(computation))
-        except BoaError as b:
-            # modify the error so the traceback starts in userland.
-            # inspired by answers in https://stackoverflow.com/q/1603940/
-            raise strip_internal_frames(b) from None
-
-    @property
-    def address(self) -> Address:
-        return self._address
 
 
 # a few lines of shared code between VyperBlueprint and VyperContract
@@ -306,43 +247,6 @@ class ErrorDetail:
                 msg += f" {self.frame_detail}"
 
         return msg
-
-
-class StackTrace(list):
-    def __str__(self):
-        return "\n\n".join(str(x) for x in self)
-
-    @property
-    def last_frame(self):
-        return self[-1]
-
-
-def trace_for_unknown_contract(computation, env):
-    ret = StackTrace(
-        [f"<Unknown location in unknown contract {computation.msg.code_address.hex()}>"]
-    )
-    return _handle_child_trace(computation, env, ret)
-
-
-def _handle_child_trace(computation, env, return_trace):
-    if len(computation.children) == 0:
-        return return_trace
-    if not computation.children[-1].is_error:
-        return return_trace
-    child = computation.children[-1]
-
-    # TODO: maybe should be:
-    # child_obj = (
-    #   env.lookup_contract(child.msg.code_address)
-    #   or env._code_registry.get(child.msg.code)
-    # )
-    child_obj = env.lookup_contract(child.msg.code_address)
-
-    if child_obj is None:
-        child_trace = trace_for_unknown_contract(child, env)
-    else:
-        child_trace = child_obj.stack_trace(child)
-    return StackTrace(child_trace + return_trace)
 
 
 # "pattern match" a BoaError. tries to match fields of the error
@@ -946,10 +850,6 @@ class VyperContract(_BaseContract):
         setattr(self.inject, fn_ast.name, f)
 
 
-class _EvmFunction:
-    ...
-
-
 class VyperFunction(_EvmFunction):
     def __init__(self, fn_ast, contract):
         super().__init__()
@@ -1103,265 +1003,6 @@ class VyperInternalFunction(VyperFunction):
         return source_map
 
 
-# a contract which we only have the ABI for.
-class ABIContract(_EvmContract):
-    def __init__(
-        self,
-        name: str,
-        functions: list["ABIFunction"],
-        events: list["EventT"],
-        address: Address,
-        created_from: Optional[Address] = None,
-        filename: Optional[str] = None,
-        env=None,
-    ):
-        super().__init__(env, filename=filename, address=address)
-        self._name = name
-        self._events = events
-        self._functions = functions
-
-        for name, functions in groupby(self._functions, key=attrgetter("name")):
-            functions = list(functions)
-            fn = functions[0] if len(functions) == 1 else ABIOverload(functions)
-            fn.contract = self
-            setattr(self, name, fn)
-
-        self._address = Address(address)
-        self.created_from = created_from
-
-        self._source_map = {"pc_pos_map": {}}  # override
-
-    @cached_property
-    def method_id_map(self):
-        ret = {}
-        for function in self._functions:
-            for abi_sig, method_id_int in function.method_ids.items():
-                method_id_bytes = method_id_int.to_bytes(4, "big")
-                assert method_id_bytes not in ret  # vyper guarantees unique method ids
-                ret[method_id_bytes] = abi_sig
-        return ret
-
-    def stack_trace(self, computation: ComputationAPI):
-        calldata_method_id = bytes(computation.msg.data[:4])
-        if calldata_method_id in self.method_id_map:
-            msg = f"  (unknown location in {self}.{self.method_id_map[calldata_method_id]})"
-        else:
-            # Method might not be specified in the ABI
-            msg = f"  (unknown method id {self}.0x{calldata_method_id.hex()})"
-
-        return_trace = StackTrace([msg])
-        return _handle_child_trace(computation, self.env, return_trace)
-
-    @property
-    def deployer(self):
-        return ABIContractFactory(self._name, self._functions, self._events)
-
-    def __repr__(self):
-        ret = f"<{self._name} interface at {self.address}>"
-
-        if self.created_from is not None:
-            ret += f" (created by {self.created_from})"
-
-        return ret
-
-    # OVERRIDE
-    @cached_property
-    def event_for(self):
-        return {e.event_id: e for e in self._events}
-
-
-# name Factory instead of Deployer because it doesn't actually do any
-# contract deployment.
-class ABIContractFactory:
-    def __init__(
-        self,
-        name: str,
-        functions: list["ABIFunction"],
-        events: list["EventT"],
-        filename: Optional[str] = None,
-    ):
-        self._name = name
-        self._functions = functions
-        self._events = events
-        self._filename = filename
-
-    @classmethod
-    def from_abi_dict(cls, abi, name: Optional[str] = None):
-        if name is None:
-            name = "<anonymous contract>"
-
-        functions = [
-            ABIFunction(item, name) for item in abi if item.get("type") == "function"
-        ]
-
-        # warn on functions with same name
-        for function_name, count in Counter(f.name for f in functions).items():
-            if count > 1:
-                warnings.warn(
-                    f"{name} overloads {function_name}! overloaded methods "
-                    "might not work correctly at this time",
-                    stacklevel=1,
-                )
-
-        events = [ABIEvent.from_abi(i) for i in abi if i.get("type") == "event"]
-
-        return cls(basename(name), functions, events, filename=name)
-
-    def at(self, address) -> ABIContract:
-        """
-        Create an ABI contract object for a deployed contract at `address`.
-        """
-        address = Address(address)
-
-        ret = ABIContract(
-            self._name, self._functions, self._events, address, self._filename
-        )
-
-        bytecode = ret.env.vm.state.get_code(address.canonical_address)
-        if not bytecode:
-            warnings.warn(
-                "requested {ret} but there is no bytecode at that address!",
-                stacklevel=2,
-            )
-
-        ret.env.register_contract(address, ret)
-
-        return ret
-
-
-class ABIFunction(_EvmFunction):
-    def __init__(self, abi: dict, contract_name: str):
-        super().__init__()
-        self._abi = abi
-        self._contract_name = contract_name
-        self._function_visibility = FunctionVisibility.EXTERNAL
-        self._mutability = StateMutability.from_abi(abi)
-        self.contract: Optional["ABIContract"] = None
-
-    @cached_property
-    def return_type(self) -> Optional[VyperType]:
-        outputs = self._abi["outputs"].copy()
-        if not outputs:
-            return None
-        types = [parse_abi_type(output) for output in outputs]
-        return types[0] if len(types) == 1 else TupleT(tuple(types))
-
-    @cached_property
-    def name(self):
-        return self._abi["name"]
-
-    @cached_property
-    def arguments(self) -> list[PositionalArg]:
-        return [
-            PositionalArg(item["name"], type_from_abi(item))
-            for item in self._abi["inputs"]
-        ]
-
-    @cached_property
-    def argument_types(self) -> list[VyperType]:
-        return [arg.typ for arg in self.arguments]
-
-    @cached_property
-    def method_ids(self) -> dict[str, int]:
-        """
-        Dict of `{signature: four byte selector}` for this function.
-
-        * For functions without default arguments the dict contains one item.
-        * For functions with default arguments, there is one key for each
-          function signature.
-        """
-        arg_types = [i.canonical_abi_type for i in self.argument_types]
-        return _generate_method_id(self.name, arg_types)
-
-    def __repr__(self) -> str:
-        arg_types = ",".join(repr(a) for a in self.argument_types)
-        return (
-            f"ABI {self._contract_name}.{self.name}({arg_types}) -> {self.return_type}"
-        )
-
-    def __str__(self) -> str:
-        return repr(self)
-
-    @property
-    def is_mutable(self) -> bool:
-        return self._mutability > StateMutability.VIEW
-
-    # OVERRIDE
-    @property
-    def func_t(self):  # TODO: Do we really need this?
-        return self
-
-    # OVERRIDE
-    @cached_property
-    def _source_map(self):
-        return {"pc_pos_map": {}}
-
-    def __eq__(self, other):
-        return isinstance(other, ABIFunction) and self._abi == other._abi
-
-    def __hash__(self):
-        return hash(self._abi)
-
-    def prepare_calldata(self, *args, **kwargs):
-        if len(kwargs) + len(args) != len(self.arguments):
-            raise Exception(
-                f"Bad args to `{repr(self)}` "
-                f"(expected {len(self.arguments)} arguments, got {len(args)})"
-            )
-        args = [getattr(arg, "address", arg) for arg in args]
-        encoded_args = abi_encode(self.signature, args)
-        return method_id(self.name + self.signature) + encoded_args
-
-    @cached_property
-    def signature(self) -> str:
-        arg_types = [i.canonical_abi_type for i in self.argument_types]
-        return f"({','.join(arg_types)})"
-
-    def __call__(self, *args, value=0, gas=None, sender=None, **kwargs):
-        if not self.contract or not self.contract.env:
-            raise Exception(f"Cannot call {self} without deploying contract.")
-
-        computation = self.contract.env.execute_code(
-            to_address=self.contract.address,
-            sender=sender,
-            data=self.prepare_calldata(*args, **kwargs),
-            value=value,
-            gas=gas,
-            is_modifying=self.is_mutable,
-            contract=self.contract,
-        )
-
-        return self.contract.marshal_to_python(computation, self.return_type)
-
-
-class ABIOverload(_EvmFunction):
-    def __init__(self, functions: list[ABIFunction]):
-        super().__init__()
-        self.functions = functions
-
-    @cached_property
-    def name(self):
-        return self.functions[0].name
-
-    @property
-    def contract(self):
-        return self.functions[0].contract
-
-    @contract.setter
-    def contract(self, value):
-        for f in self.functions:
-            f.contract = value
-
-    def __call__(self, *args, **kwargs):
-        arg_count = len(kwargs) + len(args)
-        for function in self.functions:
-            if arg_count == len(function.arguments):
-                return function(*args, **kwargs)
-        raise Exception(
-            f"Could not find matching {self.name} function for given arguments {args} {kwargs}."
-        )
-
-
 class _InjectVyperFunction(VyperFunction):
     def __init__(self, contract, fn_source):
         ast, ir_executor, bytecode, source_map, _ = compile_vyper_function(
@@ -1373,20 +1014,6 @@ class _InjectVyperFunction(VyperFunction):
         self._override_bytecode = bytecode
         self._ir_executor = ir_executor
         self._source_map = source_map
-
-
-@dataclass
-class BoaError(Exception):
-    stack_trace: StackTrace
-
-    # perf TODO: don't materialize the stack trace until we need it,
-    # i.e. BoaError ctor only takes what is necessary to construct the
-    # stack trace but does not require the actual stack trace itself.
-    def __str__(self):
-        frame = self.stack_trace.last_frame
-        err = frame.vm_error
-        err.args = (frame.pretty_vm_reason, *err.args[1:])
-        return f"{err}\n\n{self.stack_trace}"
 
 
 _typ_cache = {}
